only 1 card appears in the rv

+++ app/src/main/java/com/example/weatherapplication/Model/Minutely.kt
@@ -6,6 +6,8 @@ import com.google.gson.annotations.SerializedName
 data class Minutely (
 
   @SerializedName("dt"            ) var dt            : Int? = null,
-  @SerializedName("precipitation" ) var precipitation : Int? = null
+  //@SerializedName("precipitation" ) var precipitation : Int? = null
+  @SerializedName("precipitation" ) var precipitation : Double? = null
+
 
 )
